removed aids from utils.hpp defines que trencaven les coses i changed uints per size_t

+++ MyButton.cpp
@@ -11,7 +11,7 @@ Button::Button(){
 void Button::setTextColor(sf::Color c){text.setColor(c); }
 sf::Color Button::getTextColor(){ return text.getColor(); }
 
-void Button::setCharacterSize(uint u){ text.setCharacterSize(u); }
+void Button::setCharacterSize(size_t u){ text.setCharacterSize(u); }
 uint Button::getCharacterSize(){ return text.getCharacterSize(); }
 
 void Button::setFont(sf::Font f){ font = f; text.setFont(font); }
